fixed 25c6165 from https://gitee.com/lverpeng/drivers_peripheral/pulls/2076 告警处理，删除BUILD.gn中涉及告警的部分

+++ bluetooth/audio/hal/hdi_passthrough/include/audio_internal.h
@@ -283,7 +283,7 @@ struct AudioHwRenderParam {
 struct ErrorDump {
     int32_t errorCode;
     int32_t count;
-    uint64_t frames;
+    unsigned long long frames;
     char* reason;                       // Specific reasons for failure
     char* currentTime;
 };
